refactor booking

+++ internal/dto/current_vendor_orders_stats_response_dto.go
@@ -4,10 +4,10 @@ package dto
 // response of current vendor orders stats DTO.
 type CurrentVendorOrdersStatsResponseDTO struct {
 	// TotalOrders is the total number of orders.
-	TotalOrders int `json:"total_orders"`
+	TotalOrders int64 `json:"total_orders"`
 
 	// TotalOrdersToday is the total number of orders today.
-	TotalOrdersToday int `json:"total_orders_today"`
+	TotalOrdersToday int64 `json:"total_orders_today"`
 
 	// RecentOrders is the recent orders.
 	RecentOrders *[]CurrentVendorOrderDTO `json:"recent_orders"`
